backend/requirements.txt
asgiref==3.11.1
Django==5.2.11
djangorestframework==3.16.1
sqlparse==0.5.5
typing_extensions==4.15.0
django-cors-headers==4.3.0  # ou la dernière version compatible
psycopg2-binary==2.9.9



